+++ .gitignore
@@ -215,3 +215,5 @@ dmypy.json
 
 # Cython debug symbols
 cython_debug/
+
+data/
