+++ data/food/0005/generate_joint_cluster.sh
