+++ data/food/0005/new_xy_embeddings.csv
