+++ data/food/0005/original_xy_embeddings.csv
