+++ defaultconfig.txt
@@ -4,8 +4,8 @@ experimentID: 0000
 baseDir: ./experiment/%(dataset)s/%(experimentID)s
 generatedDir: ./data/%(dataset)s/%(experimentID)s
 metricDir: %(baseDir)s/metrics
-geojsonDir: ./%(generatedDir)s
-externalDir: ./%(generatedDir)s
+geojsonDir: %(generatedDir)s
+externalDir: %(generatedDir)s
 webCacheDir: %(baseDir)s/webCache
 
 [ExternalFiles]
@@ -15,7 +15,6 @@ names_with_id: %(externalDir)s/domain_concept.csv
 article_coordinates: %(generatedDir)s/new_xy_embeddings.csv
 zpop_with_id: %(generatedDir)s/popularity_score.csv
 clusters_with_id: %(generatedDir)s/key_phrases_cluster_groups.csv
-
 [PreprocessingConstants]
 sample_borders: True
 sample_size: 50000
@@ -45,7 +44,7 @@ base_url: http://localhost:4000/%(dataset)s
 compress_png: true
 
 [MapData]
-countries_geojson: %(geojsonDir)s/countries_new_try.geojson
+countries_geojson: %(geojsonDir)s/countries_why.geojson
 
 [MapResources]
 img_dot: ./data/labdata/blackDot.png
